// src/app/models/UserActivityViewModel.ts
export class UserActivityViewModel {
  Id: number;
  Activity: string;
  Remarks: string;
  IsActivitySucceeded: boolean;
  HostIP: string;
  HostBrowserDetails: string;
  HostTimeZone: string;
  ActivityDateTime: Date;
  ActionById: number;
  ActionForId: number;
  ProjectId: number;
  ParameterCarrier: string;
}